Rename class file name
