+++ src/game/UserManager.js
